refactor: useAnimationEnd 훅 파일 이름 변경
