+++ frontend/src/pages/CheckEvent/hooks/useAnimationEnd.ts
